refactor(admin): 자묻질 수정에도 반영
