+++ apps/admin/src/components/faq/FaqEdit/faqEdit.hooks.ts
@@ -1,5 +1,54 @@
+import { useState, useRef, useEffect, ChangeEventHandler } from 'react';
 import { usePutFaqMutation } from '@/services/faq/mutations';
-import type { FaqInput } from '@/types/faq/client';
+import { useFaqDetailQuery } from '@/services/faq/queries';
+import { resizeTextarea } from '@/utils';
+import { FaqCategory, FaqInput } from '@/types/faq/client';
+
+export const useFaqEditData = (id: number) => {
+  const { data: faqDetailData } = useFaqDetailQuery(id);
+  const contentTextareaRef = useRef<HTMLTextAreaElement>(null);
+
+  const [faqData, setFaqData] = useState<FaqInput>({
+    title: '',
+    content: '',
+    category: 'SCHOOL_LIFE',
+  });
+
+  useEffect(() => {
+    if (faqDetailData) {
+      setFaqData({
+        title: faqDetailData.title,
+        content: faqDetailData.content,
+        category: faqDetailData.category,
+      });
+    }
+  }, [faqDetailData]);
+
+  const handleFaqDataChange: ChangeEventHandler<
+    HTMLInputElement | HTMLTextAreaElement
+  > = (e) => {
+    const { name, value } = e.target;
+    setFaqData((prev) => ({ ...prev, [name]: value }));
+
+    if (name === 'content') {
+      resizeTextarea(contentTextareaRef);
+    }
+  };
+
+  const handleFaqCategoryChange = (value: string, name: string) => {
+    setFaqData((prev) => ({ ...prev, category: value as FaqCategory }));
+  };
+
+  useEffect(() => resizeTextarea(contentTextareaRef), []);
+
+  return {
+    faqData,
+    setFaqData,
+    contentTextareaRef,
+    handleFaqDataChange,
+    handleFaqCategoryChange,
+  };
+};
 
 export const useFaqEditAction = (id: number, faqData: FaqInput) => {
   const { putFaqMutate } = usePutFaqMutation(id);
